third_party/blink/web_tests/animations/animationworklet/worklet-animation-currentTime.html

<!DOCTYPE html>
<title>Test that opacity changes when animation starts</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
#target {
  width: 100px;
  height: 100px;
  background-color: green;
}
</style>

<script id="simple_animate" type="text/worklet">
registerAnimator("test_animator", class {
  animate(currentTime, effect) {
    effect.localTime = currentTime;
  }
});
</script>
<div id="target"></div>
<script src="resources/animation-worklet-tests.js"></script>
<script>
promise_test(async t => {
  await runInAnimationWorklet(document.getElementById('simple_animate').textContent);
  const effect = new KeyframeEffect(target, [{ opacity: 0 }], { duration: 1000 });
  const animation = new WorkletAnimation('test_animator', effect);
  const first_opacity = getComputedStyle(target).opacity;
  assert_equals(first_opacity, '1');
  animation.play();
  // Wait for one async animation frame to ensure animation is running.
  await waitForAsyncAnimationFrame();
  await waitForDocumentTimelineAdvance();
  assert_greater_than(animation.currentTime, 0);
  // At this point we are guaranteed that opacity value is updated in the
  // worklet but  getComputedStyle needs an extra frame to fetch the updated
  // value to main thread.
  await new Promise(window.requestAnimationFrame);
  const second_opacity = getComputedStyle(target).opacity;
  assert_true(second_opacity < first_opacity);
}, 'Opacity should change as the animation starts.');
</script>
